+++ src/simulation/Simulation.cpp
@@ -162,7 +162,7 @@ Simulation::crankAllNodes(int nbTicks)
 
 bool Simulation::haveAllExternalized(int num)
 {
-    uint64_t min = INT_MAX;
+    uint32_t min = UINT_MAX;
     for(auto it = mNodes.begin(); it != mNodes.end(); ++it) 
     {
         auto n = it->second->getLedgerMaster().getLedgerNum();
